% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extracts}
\alias{extracts}
\title{Extract matching text from progress notes; multiple patterns allowed}
\usage{
extracts(
  pats,
  dat_pn,
  varname_pn = "PROGRESS_NOTE",
  varname_enc_id = "PAT_ENC_CSN_ID",
  varname_pat_id = "PAT_ID",
  varname_pat_mrn = "PAT_MRN",
  varname_date = "CONTACT_DATE",
  verbose = 0
)
}
\arguments{
\item{pats}{vector of regular expressions for which to search}

\item{dat_pn}{data.table or data.frame containing clinic visit notes}

\item{varname_pn}{name of variable in dat_pn containing the progress note.  Default is "PROGRESS_NOTE".}

\item{varname_enc_id}{name of variable in dat_pn containing the progress note.  Default is "PAT_ENC_CSN_ID".}

\item{varname_pat_id}{name of variable in dat_pn containing the progress note.  Default is "PAT_ID".}

\item{varname_pat_mrn}{name of variable in dat_pn containing the progress note.  Default is "PAT_MRN".}

\item{varname_date}{name of variable in dat_pn containing the progress note.  Default is "CONTACT_DATE".}

\item{verbose}{Controls the amount of output to the console. The default, 0, prints nothing.  Higher values provide more detail.}
}
\value{
data.frame (invisibly) with seven columns containing all text matching pattern pat, the position of the matches within the progress note and the associated data for that progress note (encounter, mrn, and id tokens; date). The variable names of the output are always "PAT_ENC_CSN_ID", "PAT_ID", "PAT_MRN", "CONTACT_DATE", "MATCHED_TEXT", "POSITION", and "LINE", regardless of the input variable names.
}
\description{
Extract matching text from progress notes; multiple patterns allowed
}
\examples{
dat_pn <- data.frame(
  PAT_MRN = c("A", "B"),
  PAT_ID = c("ZA", "ZB"),
  CONTACT_DATE = c("2023-03-01", "2020-03-01"),
  PAT_ENC_CSN_ID = c("A1", "B1"),
  PROGRESS_NOTE = c("XLRP uh oh", "No XLRP. Good day."))
system.time(dat_ext <- extracts(
  pats = c(
    "XLRP",
    "x[ -]*linked[ -]*(RP|retinitis([ -]*pigmentosa)?)"),
  dat_pn))
table(dat_ext$MATCHED_TEXT)
}
