% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negationprecautionhistory.R
\name{findprecautions}
\alias{findprecautions}
\title{find precautions terms in progress notes}
\usage{
findprecautions(dat_pn, varname_pn = "PROGRESS_NOTE", verbose = 0)
}
\arguments{
\item{dat_pn}{data.frame containing ENC_VISIT data}

\item{varname_pn}{name of PROGRESS_NOTE in dat_pn}

\item{verbose}{Controls the amount of output to the console. The default, 0, prints nothing.  Higher values provide more detail.}
}
\value{
data.frame of precaution terms with information about location.
}
\description{
find precautions terms in progress notes
}
\examples{
# findprecautions()
}
