% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{keyword2encounter}
\alias{keyword2encounter}
\title{Title aggregate collection of keyword data.frame to patient level}
\usage{
keyword2encounter(dat_enc_neg)
}
\arguments{
\item{dat_enc_neg}{data.frame}
}
\value{
data.frame with variables for mrn token, encounter token, contact date, n_positive, n_negative
}
\description{
Title aggregate collection of keyword data.frame to patient level
}
