% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{keyword2patient}
\alias{keyword2patient}
\title{Title aggregate collection of keyword data.frame to patient level}
\usage{
keyword2patient(dat_enc_neg)
}
\arguments{
\item{dat_enc_neg}{data.frame}
}
\value{
data.frame with variables for mrn, n_positive, n_negative, first date, last date, n dates
}
\description{
Title aggregate collection of keyword data.frame to patient level
}
