% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extracts_with_exceptions}
\alias{extracts_with_exceptions}
\title{Extract matching text from progress notes; multiple patterns allowed and exclusion patterns allowed}
\usage{
extracts_with_exceptions(
  yespats,
  nopats,
  dat_pn,
  varname_pn = "PROGRESS_NOTE",
  verbose = 0
)
}
\arguments{
\item{yespats}{vector of regular expressions for which to search}

\item{nopats}{vector of regular expressions for which to exclude}

\item{dat_pn}{data.table or data.frame containing clinic visit notes}

\item{varname_pn}{name of variable in dat_pn containing the progress note.  Default is "PROGRESS_NOTE".}

\item{verbose}{Controls the amount of output to the console. The default, 0, prints nothing.  Higher values provide more detail.}
}
\value{
data.frame or data.table (invisibly) with seven columns containing all text matching pattern pat, the position of the matches within the progress note and the associated data for that progress note (encounter, mrn, and id tokens; date)
}
\description{
Extract matching text from progress notes; multiple patterns allowed and exclusion patterns allowed
}
\examples{
dat_pn <- data.frame(
  PAT_MRN = c("A", "B"),
  PAT_ID = c("ZA", "ZB"),
  CONTACT_DATE = c("2023-03-01", "2020-03-01"),
  PAT_ENC_CSN_ID = c("A1", "B1"),
  PROGRESS_NOTE = c("XLRP uh oh", "No XLRP. Good day."))
extracts_with_exceptions(
   yespats = c("XLRP"),
   nopats = c("\\\\bRP2\\\\b"),
   dat_pn)
}
