% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadprogressnotes.R
\name{loadprogressnotes}
\alias{loadprogressnotes}
\title{Load Progress Notes: This function reads the ENC_VISIT file in preparation for free text search of the progress note.}
\usage{
loadprogressnotes(
  filename,
  varname_enc_id = "PAT_ENC_CSN_ID",
  verbose = 0,
  usefread = TRUE,
  nrows = NULL
)
}
\arguments{
\item{filename}{Name of encounter visit file.  At UM this is "SOURCE_UM_OPH_ENC_VISIT.csv".  You may wish to include entire path to the file.}

\item{varname_enc_id}{Name of the encounter id variable filename.  At UM this is "PAT_ENC_CSN_ID" (which is the default if nothing else is provided).}

\item{verbose}{Controls the amount of output to the console. The default, 0, prints nothing.  Higher values provide more detail.}

\item{usefread}{If TRUE (default), use data.table::fread. If FALSE, use utils::read.csv.}

\item{nrows}{Number of rows to read from file.  Default (-1) is to read all rows.  Consider positive values for testing.}
}
\value{
A data.table (fread == TRUE) or data.frame (fread == FALSE), invisibly.
}
\description{
Load Progress Notes: This function reads the ENC_VISIT file in preparation for free text search of the progress note.
}
\examples{
## Not Run
## dat_pn <- loadprogressnotes("K:/lab/Scrubber/ToScrub/2021/SOURCE_UM_OPH_ENC_VISIT.csv")
}
